Added stream_line() function  ... but I have concerns:
If a file is meant to have a \r at the end of a line (using \n terminators) it will
lose that character (we will see it as part of a \r\n terminator).

I think it is better to specify which terminator you expect, in which case this
function isn't necessary (stream until token is probably sufficient).  In the case
of HTTP multipart, the terminator after the first boundary should be remembered and
used as the terminator subsequently sought.