Removed async code: that code never worked:
It must have been a low-coffee day when I wrote the async code. It just calls the
sync code, and expects the underlying I/O system to report WouldBlock events, which
it never does because it never calls the underlying I/O with flags that might get
such a result.

I think I knew this when I wrote it, but then I forgot where I left things and when
I came back to it I thought it must have been finished already.

I apologize if anyone has had async troubles from using it.